refactored the business page code
